pyproject.toml
[project]
name = "transcribe"
version = "1.0.0"
description = "Voice transcription app using Nvidia Parakeet model with global hotkey"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "nemo_toolkit[asr]",
    "sounddevice",
    "numpy",
    "scipy",
    "python-xlib",
]

[project.scripts]
transcribe = "transcribe.app:main"

[tool.commitizen]
version = "1.0.0"
version_files = [
    "__version__.py",
    "pyproject.toml:version",
]
update_changelog_on_bump = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/transcribe"]

[tool.transcribe]
# Model choices:
#   "nvidia/parakeet-tdt-0.6b-v3"   - fast, lightweight (default)
#   "nvidia/parakeet-rnnt-1.1b"     - slower, higher accuracy
model = "nvidia/parakeet-tdt-0.6b-v3"

# Hotkey as modifier+key string. Modifiers: ctrl, shift, alt, super
# Key: any single character or X11 key name
hotkey = "ctrl+shift+;"

[tool.ruff]
line-length = 79

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[dependency-groups]
dev = [
    "pytest",
    "pytest-mock",
    "ruff",
    "pre-commit",
]
